# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: PESATRIX Open Education Infrastructure API
  description: Reusable, secure, interoperable digital infrastructure for structured
    educational events and information in low-resource environments. No personal data
    is stored by this system.
  version: 0.1.0
paths:
...
  /api/v1/users:
    post:
      tags:
      - users
      summary: Create a user (admin)
      operationId: create_api_v1_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - users
      summary: List users (admin)
      operationId: list_users_api_v1_users_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_UserResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations:
    post:
      tags:
      - organizations
      summary: Create an organization
      operationId: create_org_api_v1_organizations_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - organizations
      summary: List organizations
      operationId: list_orgs_api_v1_organizations_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_OrganizationResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{organization_id}:
    get:
      tags:
      - organizations
      summary: Get an organization
      operationId: get_org_api_v1_organizations__organization_id__get
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Organization Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - organizations
      summary: Update an organization (admin)
      operationId: update_org_api_v1_organizations__organization_id__patch
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Organization Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{organization_id}/schools:
    post:
      tags:
      - organizations
      summary: Create a school under an organization
      operationId: create_school_route_api_v1_organizations__organization_id__schools_post
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Organization Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - organizations
      summary: List schools in an organization
      operationId: list_schools_route_api_v1_organizations__organization_id__schools_get
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Organization Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_SchoolResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/schools/{school_id}:
    get:
      tags:
      - organizations
      summary: Get a school
      operationId: get_school_route_api_v1_organizations_schools__school_id__get
      parameters:
      - name: school_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: School Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/sync/devices:
    post:
      tags:
      - sync
      summary: Register a device for offline sync
      operationId: register_api_v1_sync_devices_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync/upload:
    post:
      tags:
      - sync
      summary: Upload an offline event batch
      operationId: upload_api_v1_sync_upload_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncUploadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncUploadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/sync/download:
    post:
      tags:
      - sync
      summary: Download events since a watermark
      operationId: download_api_v1_sync_download_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncDownloadRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncDownloadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DeviceCreate:
      properties:
        school_id:
          type: string
          format: uuid
          title: School Id
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        device_type:
          type: string
          title: Device Type
          default: mobile
      type: object
      required:
      - school_id
      - name
      title: DeviceCreate
    DeviceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        school_id:
          type: string
          format: uuid
          title: School Id
        name:
          type: string
          title: Name
        device_type:
          type: string
          title: Device Type
        status:
          $ref: '#/components/schemas/DeviceStatus'
        last_seen_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen At
      type: object
      required:
      - id
      - created_at
      - updated_at
      - school_id
      - name
      - device_type
      - status
      title: DeviceResponse
    DeviceStatus:
      type: string
      enum:
      - active
      - suspended
      - retired
      title: DeviceStatus
...
    EventKind:
      type: string
      enum:
      - check_in
      - check_out
      - lesson_started
      - lesson_completed
      - attendance_override
      - credential_issued
      - credential_verified
      - sync_batch
      title: EventKind
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrganizationCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        code:
          type: string
          maxLength: 64
          minLength: 2
          pattern: ^[a-zA-Z0-9_-]+$
          title: Code
        org_type:
          $ref: '#/components/schemas/OrganizationType'
          default: other
        parent_organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Organization Id
      type: object
      required:
      - name
      - code
      title: OrganizationCreate
    OrganizationResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        name:
          type: string
          title: Name
        code:
          type: string
          title: Code
        org_type:
          $ref: '#/components/schemas/OrganizationType'
        status:
          $ref: '#/components/schemas/RecordStatus'
        parent_organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Organization Id
      type: object
      required:
      - id
      - created_at
      - updated_at
      - name
      - code
      - org_type
      - status
      title: OrganizationResponse
    OrganizationType:
      type: string
      enum:
      - school
      - district
      - nonprofit
      - government
      - other
      title: OrganizationType
    OrganizationUpdate:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Name
      type: object
      title: OrganizationUpdate
...
    PaginatedResponse_OrganizationResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OrganizationResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[OrganizationResponse]
    PaginatedResponse_SchoolResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SchoolResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[SchoolResponse]
    PaginatedResponse_UserResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - items
      - total
      - page
      - size
      - has_more
      title: PaginatedResponse[UserResponse]
    RecordStatus:
      type: string
      enum:
      - active
      - suspended
      - archived
      title: RecordStatus
...
    SchoolCreate:
      properties:
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        code:
          type: string
          maxLength: 64
          minLength: 2
          pattern: ^[a-zA-Z0-9_-]+$
          title: Code
      type: object
      required:
      - organization_id
      - name
      - code
      title: SchoolCreate
    SchoolResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        name:
          type: string
          title: Name
        code:
          type: string
          title: Code
        status:
          $ref: '#/components/schemas/RecordStatus'
      type: object
      required:
      - id
      - created_at
      - updated_at
      - organization_id
      - name
      - code
      - status
      title: SchoolResponse
    SyncDownloadRequest:
      properties:
        device_id:
          type: string
          format: uuid
          title: Device Id
        since_seq:
          type: integer
          title: Since Seq
          default: 0
        limit:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          default: 200
      type: object
      required:
      - device_id
      title: SyncDownloadRequest
    SyncDownloadResponse:
      properties:
        device_id:
          type: string
          format: uuid
          title: Device Id
        since_seq:
          type: integer
          title: Since Seq
        watermark_after:
          type: integer
          title: Watermark After
        records:
          items:
            $ref: '#/components/schemas/EducationalEventResponse'
          type: array
          title: Records
        more:
          type: boolean
          title: More
      type: object
      required:
      - device_id
      - since_seq
      - watermark_after
      - records
      - more
      title: SyncDownloadResponse
    SyncUploadRecord:
      properties:
        idempotency_key:
          type: string
          maxLength: 128
          minLength: 8
          pattern: ^[\w\-]+$
          title: Idempotency Key
        school_id:
          type: string
          format: uuid
          title: School Id
        student_identity_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Student Identity Id
        event_kind:
          $ref: '#/components/schemas/EventKind'
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        device_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Device Id
        source:
          type: string
          maxLength: 24
          title: Source
          default: offline_device
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - idempotency_key
      - school_id
      - event_kind
      - occurred_at
      title: SyncUploadRecord
    SyncUploadRequest:
      properties:
        device_id:
          type: string
          format: uuid
          title: Device Id
        batch_seq:
          type: integer
          minimum: 1.0
          title: Batch Seq
        records:
          items:
            $ref: '#/components/schemas/SyncUploadRecord'
          type: array
          minItems: 1
          title: Records
      type: object
      required:
      - device_id
      - batch_seq
      - records
      title: SyncUploadRequest
    SyncUploadResponse:
      properties:
        accepted:
          type: integer
          title: Accepted
        duplicates:
          type: integer
          title: Duplicates
        errors:
          type: integer
          title: Errors
          default: 0
        watermark_after:
          type: integer
          title: Watermark After
        batch_id:
          type: string
          format: uuid
          title: Batch Id
        failed:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Failed
      type: object
      required:
      - accepted
      - duplicates
      - watermark_after
      - batch_id
      title: SyncUploadResponse
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: ^[a-zA-Z0-9_.-]+$
          title: Username
        password:
          type: string
          maxLength: 256
          minLength: 8
          title: Password
        display_name:
          type: string
          maxLength: 120
          minLength: 1
          title: Display Name
        role:
          type: string
          title: Role
          default: operator
        organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organization Id
        school_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: School Id
      type: object
      required:
      - username
      - password
      - display_name
      title: UserCreate
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
        role:
          type: string
          title: Role
        status:
          $ref: '#/components/schemas/RecordStatus'
        organization_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organization Id
        school_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: School Id
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
      type: object
      required:
      - id
      - created_at
      - updated_at
      - username
      - display_name
      - role
      - status
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError